requirements.txt
flask
psycopg2-binary
waitress
nltk
